Extract boolean validator into dedicated Trait
In order to ease Unit Test redability
