+++ src/UI/Resolver/Validator/RawValueValidator.php
@@ -8,6 +8,7 @@ use Assert\InvalidArgumentException;
 use Imedia\Ammit\Domain\DateValidation;
 use Imedia\Ammit\UI\Resolver\Exception\UIValidationException;
 use Imedia\Ammit\UI\Resolver\UIValidationEngine;
+use Imedia\Ammit\UI\Resolver\Validator\Implementation\Pure\BooleanValidatorTrait;
 use Imedia\Ammit\UI\Resolver\Validator\Implementation\Pure\IntegerValidatorTrait;
 
 /**
@@ -15,6 +16,7 @@ use Imedia\Ammit\UI\Resolver\Validator\Implementation\Pure\IntegerValidatorTrait
  */
 class RawValueValidator implements UIValidatorInterface
 {
+    use BooleanValidatorTrait;
     use IntegerValidatorTrait;
 
     /** @var UIValidationEngine */
@@ -45,34 +47,6 @@ class RawValueValidator implements UIValidatorInterface
         );
 
         return $value;
-    }
-
-    /**
-     * Exceptions are caught in order to be processed later
-     * @param mixed $value Boolean ?
-     *
-     * @return boolean Value casted into boolean or false
-     */
-    public function mustBeBoolean($value, string $propertyPath = null, UIValidatorInterface $parentValidator = null, string $exceptionMessage = null): bool
-    {
-        $this->validationEngine->validateFieldValue(
-            $parentValidator ?: $this,
-            function() use ($value, $propertyPath, $exceptionMessage) {
-                Assertion::inArray(
-                    $value,
-                    [true, false, 1, 0, '1', '0', 'true', 'false'],
-                    $exceptionMessage,
-                    $propertyPath
-                );
-            }
-        );
-
-        // Otherwise "false" would return true
-        if (in_array($value, [true, 'true', '1', 1], true)) {
-            return true;
-        }
-
-        return false;
     }
 
     /**
