+++ tests/units/UI/Resolver/Validator/RawValueValidator.php
@@ -89,108 +89,6 @@ class RawValueValidator extends atoum
     }
 
     /**
-     * @dataProvider booleanDataProvider
-     */
-    public function test_it_gets_value_even_if_boolean_valid($value, $expected)
-    {
-        // Given
-        $uiValidationEngine = UIValidationEngine::initialize();
-        $sut = new SUT($uiValidationEngine);
-
-        // When
-        $actual = $sut->mustBeBoolean(
-            $value,
-            'accept',
-            null,
-            'Custom Exception message'
-        );
-
-        // Then
-        $this
-            ->variable($actual)
-            ->isEqualTo($expected);
-
-        $uiValidationEngine->guardAgainstAnyUIValidationException();
-    }
-
-    protected function booleanDataProvider(): array
-    {
-        return [
-            [
-                'value' => true,
-                'expected' => true
-            ],
-            [
-                'value' => false,
-                'expected' => false
-            ],
-            [
-                'value' => 'true',
-                'expected' => true
-            ],
-            [
-                'value' => 'false',
-                'expected' => false
-            ],
-            [
-                'value' => 1,
-                'expected' => true
-            ],
-            [
-                'value' => 0,
-                'expected' => false
-            ],
-        ];
-    }
-
-    public function test_it_gets_value_even_if_bad_boolean_detected()
-    {
-        // Given
-        $value = 'bad';
-        $propertyPath = 'latitude';
-        $errorMessage = 'Custom Exception message';
-
-        $expectedNormalizedExceptions = [
-            'errors' => [
-                [
-                'status' => 406,
-                'source' => ['parameter' => $propertyPath],
-                'title' => 'Invalid Parameter',
-                'detail' => $errorMessage,
-                ]
-            ]
-        ];
-
-        $uiValidationEngine = UIValidationEngine::initialize();
-        $sut = new SUT($uiValidationEngine);
-
-        // When
-        $actual = $sut->mustBeBoolean(
-            $value,
-            $propertyPath,
-            null,
-            $errorMessage
-        );
-
-        // Then
-        $this
-            ->variable($actual)
-            ->isEqualTo(false);
-
-        try {
-            $uiValidationEngine->guardAgainstAnyUIValidationException();
-        } catch (UIValidationCollectionException $e) {
-            $actual = $e->normalize();
-            $this->array($actual)
-                ->isEqualTo($expectedNormalizedExceptions);
-
-            return;
-        }
-
-        $this->throwError();
-    }
-
-    /**
      * @dataProvider notArrayDataProvider
      */
     public function test_it_gets_value_even_if_not_array_value_detected($propertyPath, $errorMessage, $value, array $expectedNormalizedException)
